requirements/base.txt
aws-cdk.core==1.79.0
aws-cdk.aws-lambda==1.79.0
aws-cdk.aws-s3==1.79.0
aws-cdk.aws-ecr==1.79.0
aws-cdk.aws-rds==1.79.0
aws-cdk.aws-codebuild==1.79.0
aws-cdk.aws-codedeploy==1.79.0
aws-cdk.aws-codepipeline==1.79.0
aws-cdk.aws-codepipeline-actions==1.79.0
aws-cdk.aws-elasticloadbalancingv2==1.79.0

boto3==1.16.35
rich==9.5.1
invoke==1.4.1
